[Debug/debug] 디버그, 일부기능 수정

+++ app/src/main/java/com/devik/homebarorder/data/source/local/database/MenuEntity.kt
@@ -9,7 +9,7 @@ data class MenuEntity(
     @PrimaryKey(autoGenerate = true) val uid: Int = 0,
     val menuName: String,
     val menuInfo: String = "",
-    val menuPrice: Int = 0,
+    val menuPrice: Long = 0,
     val menuImage: Bitmap? = null,
     val menuCategory: Int
 )
